refactor: - libs moved inside lib/ folder - fixed nav pid implementation errors
