+++ lib/comms/include/comms.h
@@ -5,7 +5,7 @@
 #include <WiFi.h>
 #include <WebServer.h>
 #include <ElegantOTA.h>
-#include <float_common.h>
+#include "float_common.h"
 
 /*
  *******************************************************************************
