+++ lib/led/src/led.cpp
@@ -1,5 +1,5 @@
 #include "led.h"
-#include "../config.h"
+#include "config.h"
 
 /*
  *******************************************************************************
