+++ lib/pid/include/pid.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Arduino.h>e
+#include <Arduino.h>
 
 /*
  *******************************************************************************
