+++ platformio.ini
@@ -22,6 +22,7 @@ lib_deps =
 build_flags = 
     -DCORE_DEBUG_LEVEL=3
     -DCONFIG_ARDUHAL_LOG_COLORS
+    -I include
 
 ; ESPA - Float board (main controller)
 [env:espA]
@@ -32,11 +33,14 @@ upload_speed = ${esp32_common.upload_speed}
 monitor_speed = ${esp32_common.monitor_speed}
 monitor_filters = ${esp32_common.monitor_filters}
 build_src_filter = +<espA/>
-build_flags = 
+
+test_framework = unity
+test_build_src = yes
+
+build_flags =
     ${esp32_common.build_flags}
     -DESPA_BUILD
 
-; Additional libs specific to espA
 lib_deps = 
     ${esp32_common.lib_deps}
     ESPAsyncWebServer
